filterTasksByTaxonomyTerms
