+++ src/selectors/tasks/index.ts
@@ -1,8 +1,6 @@
 import * as R from 'ramda';
 import { Store } from '../../stores';
 import * as store from '../../stores/tasks';
-import { Taxonomies as TaxonomyConstants } from '../../application/constants';
-import { TaxonomyTermReference } from '../taxonomies/pull_explore_taxonomy';
 import { TaskListItem } from './task_list_item';
 import { selectTaskAsListItem } from './select_task_as_list_item';
 
@@ -19,21 +17,3 @@ export const rejectTasksWithIdsInList =
     });
 
 export const rejectCompletedTasks = R.reject(R.prop('completed'));
-
-export const filterTasksByTaxonomyTerms =
-    R.curry((selectedAnswerTaxonomyTerms: ReadonlyArray<TaxonomyTermReference>, taskMap: store.TaskMap): ReadonlyArray<store.Task> => {
-
-        const taskContainsTaxonomyTerms = R.curry((targetTerms: ReadonlyArray<TaxonomyTermReference>, task: store.Task): boolean => (
-            R.not(R.isEmpty(R.intersection(targetTerms, task.taxonomyTerms)))
-        ));
-
-        const taxonomyId = TaxonomyConstants.RECOMMENDATION_TAXONOMY_ID;
-        const taxonomyTermId = TaxonomyConstants.RECOMMEND_TO_ALL_TAXONOMY_TERM_ID;
-
-        const isRecommendedToAll = taskContainsTaxonomyTerms([{ taxonomyId, taxonomyTermId }]);
-        const isRecommendedBecauseOfSelecedAnswers = taskContainsTaxonomyTerms(selectedAnswerTaxonomyTerms);
-
-        const isRecommended = R.either(isRecommendedToAll, isRecommendedBecauseOfSelecedAnswers);
-
-        return R.filter(isRecommended, R.values(taskMap));
-    });
